neopantry/requirements/base.txt
aloe-django==0.1.3
Django==2.0.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-grappelli==2.11.1
django-localflavor==2.0
django-phonenumber-field==2.0.0
django-tinymce==2.7.0
django-wysiwyg==0.8.0
djangorestframework==3.7.7
factory-boy==2.10.0
Pillow==5.0.0
psycopg2==2.7.3.2
PyYAML==3.12
requests==2.18.4
